refactor for windows support
