+++ mlos_bench/mlos_bench/tests/services/remote/ssh/test_ssh_fileshare.py
@@ -6,12 +6,10 @@
 Tests for mlos_bench.services.remote.ssh.ssh_services
 """
 
-from logging import warning
 from os.path import basename
 from pathlib import Path
 from typing import Dict, List
 
-import filecmp
 import os
 import tempfile
 
@@ -24,7 +22,7 @@ from mlos_bench.services.remote.ssh.ssh_host_service import SshHostService
 from mlos_bench.services.remote.ssh.ssh_fileshare import SshFileShareService
 from mlos_bench.util import path_join
 
-from mlos_bench.tests import requires_docker
+from mlos_bench.tests import are_dir_trees_equal, requires_docker
 from mlos_bench.tests.services.remote.ssh import SshTestServerInfo
 
 
@@ -42,54 +40,39 @@ def test_ssh_fileshare_single_file(ssh_test_server: SshTestServerInfo,
     ]
     lines = [line + "\n" for line in lines]
 
-    with tempfile.NamedTemporaryFile(mode='w+t', encoding='utf-8') as temp_file:
+    # 1. Write a local file and upload it.
+
+    # NOTE: Since Windows doesn't allow us to reopen the file while it's still
+    # open we need to handle deletion ourselves.
+    # pylint: disable=consider-using-with
+    try:
+        temp_file = tempfile.NamedTemporaryFile(mode='w+t', encoding='utf-8', delete=False)
         temp_file.writelines(lines)
         temp_file.flush()
+        temp_file.close()
+
         ssh_fileshare_service.upload(
             params=config,
             local_path=temp_file.name,
             remote_path=remote_file_path,
         )
+        os.unlink(temp_file.name)
 
-    with tempfile.NamedTemporaryFile() as temp_file:
+        # 2. Download the remote file and compare the contents.
+
+        temp_file = tempfile.NamedTemporaryFile(mode='w+t', encoding='utf-8', delete=False)
+        temp_file.close()
         ssh_fileshare_service.download(
             params=config,
             remote_path=remote_file_path,
             local_path=temp_file.name,
         )
         # Download will replace the inode at that name, so we need to reopen the file.
-        with open(temp_file.name, mode='r', encoding='utf-8') as reopened_temp_file:
-            read_lines = reopened_temp_file.readlines()
+        with open(temp_file.name, mode='r', encoding='utf-8') as temp_file_h:
+            read_lines = temp_file_h.readlines()
             assert read_lines == lines
-
-
-def are_dir_trees_equal(dir1: str, dir2: str) -> bool:
-    """
-    Compare two directories recursively. Files in each directory are
-    assumed to be equal if their names and contents are equal.
-
-    @param dir1: First directory path
-    @param dir2: Second directory path
-
-    @return: True if the directory trees are the same and
-        there were no errors while accessing the directories or files,
-        False otherwise.
-    """
-    # See Also: https://stackoverflow.com/a/6681395
-    dirs_cmp = filecmp.dircmp(dir1, dir2)
-    if len(dirs_cmp.left_only) > 0 or len(dirs_cmp.right_only) > 0 or len(dirs_cmp.funny_files) > 0:
-        warning(f"Found differences in dir trees {dir1}, {dir2}:\n{dirs_cmp.diff_files}\n{dirs_cmp.funny_files}")
-        return False
-    (_, mismatch, errors) = filecmp.cmpfiles(dir1, dir2, dirs_cmp.common_files, shallow=False)
-    if len(mismatch) > 0 or len(errors) > 0:
-        warning(f"Found differences in files:\n{mismatch}\n{errors}")
-        return False
-    for common_dir in dirs_cmp.common_dirs:
-        new_dir1 = os.path.join(dir1, common_dir)
-        new_dir2 = os.path.join(dir2, common_dir)
-        if not are_dir_trees_equal(new_dir1, new_dir2):
-            return False
-    return True
+    finally:
+        os.unlink(temp_file.name)
 
 
 @pytest.mark.xdist_group("ssh_test_server")
@@ -155,18 +138,27 @@ def test_ssh_fileshare_download_file_dne(ssh_test_server: SshTestServerInfo,
     config = ssh_test_server.to_ssh_service_config()
 
     canary_str = "canary"
-    with tempfile.NamedTemporaryFile(mode='w+t', encoding='utf-8') as temp_file:
+
+    try:
+        # Windows doesn't allow us to reopen the file while it's still open so we
+        # have to reopen it and handle deletion ourselves.
+        # pylint: disable=consider-using-with
+        temp_file = tempfile.NamedTemporaryFile(mode='w+t', encoding='utf-8', delete=False)
         temp_file.writelines([canary_str])
         temp_file.flush()
+        temp_file.close()
+
         with pytest.raises(SFTPError):
             ssh_fileshare_service.download(
                 params=config,
                 remote_path="/tmp/file-dne.txt",
                 local_path=temp_file.name,
             )
-        with open(temp_file.name, mode='r', encoding='utf-8') as reopened_temp_file:
-            read_lines = reopened_temp_file.readlines()
-            assert read_lines == [canary_str]
+        with open(temp_file.name, mode='r', encoding='utf-8') as temp_file_h:
+            read_lines = temp_file_h.readlines()
+        assert read_lines == [canary_str]
+    finally:
+        os.unlink(temp_file.name)
 
 
 @pytest.mark.xdist_group("ssh_test_server")
